rename #27: Config 패키지 이동
